requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
faiss-cpu
numpy
